nucliadb_search/requirements.txt
starlette-prometheus
uvicorn
opentelemetry-instrumentation-fastapi
opentelemetry-instrumentation-httpx
opentelemetry-propagator-b3
fastapi-versioning==0.10.0
fastapi==0.75.0
pydantic==1.9.0
sentry-sdk==1.4.1
pyjwt==2.3.0
mmh3==3.0.0
httpx==0.19.0
aiofiles==0.8.0
types-pkg-resources==0.1.3
grpc-stubs==1.24.7
orjson==3.6.7
opentelemetry-semantic-conventions==0.28b1
aiodns==3.0.0
